+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## master (XX.XX.XXXX)
+* Feature: Allow creation of Servers with Networks
+
 ## 2.1.0 (04.05.2020)
 * Bugfix: Fix wrong Guzzle Client Type on Server, Volume and FloatingIp Model.
 * Feature: Allow overriding of Guzzle Client configuration on `LKDev\HetznerCloud\Clients\GuzzleClient`